[Silver IV] Title: solved.ac, Time: 80 ms, Memory: 5108 KB -BaekjoonHub

+++ C++17/백준/Silver/18110. solved．ac/solved．ac.cc
@@ -25,7 +25,7 @@ int main(void) {
 	for (int i = average; i < n - average; i++) {
 		res += v[i];
 	}
-	if(n) cout << static_cast<int>(round(res / (n - (average * 2))));
+	if(n) cout << round(res / (n - (average * 2)));
 	else {
 		cout << 0;
 	}
